pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whisper-toggle"
version = "1.0.0"
authors = [
    {name = "Brad Johnson", email = "brad@bradjohnson.ai"},
]
description = "Real-time voice transcription with hardware toggle button support"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: POSIX :: Linux",
    "Environment :: X11 Applications :: Gnome",
]
keywords = ["whisper", "speech-recognition", "transcription", "voice-to-text", "wayland"]
dependencies = [
    "faster-whisper>=0.10.0",
    "pyaudio>=0.2.11",
    "evdev>=1.6.0",
    "scipy>=1.10.0",
    "numpy>=1.24.0",
    "PyGObject>=3.42.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/bradjohnson/whisper-toggle"
Repository = "https://github.com/bradjohnson/whisper-toggle.git"
Issues = "https://github.com/bradjohnson/whisper-toggle/issues"
Documentation = "https://github.com/bradjohnson/whisper-toggle#readme"

[project.scripts]
whisper-toggle = "whisper_toggle.app:main"
whisper-toggle-gui = "whisper_toggle.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["whisper_toggle*"]
exclude = ["tests*"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true